backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-JWT-Extended==4.5.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
mysql-connector-python==8.1.0
PyJWT==2.12.1
werkzeug==2.3.7
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
pyotp==2.9.0
qrcode==7.4.2
